[UPDATE]配列化が必要な項目をまとめ、存在しない可能性がある項目は無視できるように変更

+++ batch/src/main/java/com/ddbj/ld/service/RegisterService.java
@@ -63,7 +63,9 @@ public class RegisterService {
         File bioProjectDir = new File(bioProjectPath);
         List<File> bioProjectFileList = Arrays.asList(Objects.requireNonNull(bioProjectDir.listFiles()));
 
-        searchModule.deleteIndex(hostname, port, scheme, bioProjectIndexName);
+        if(searchModule.existsIndex(hostname, port, scheme, bioProjectIndexName)) {
+            searchModule.deleteIndex(hostname, port, scheme, bioProjectIndexName);
+        }
 
         for(File bioProjectFile: bioProjectFileList) {
             Map<String, String> bioProjectJsonMap = bioProjectParser.parse(bioProjectFile.getAbsolutePath());
